Update room booking flow
