+++ src/components/Hotels/HotelsStyles.js
@@ -1,5 +1,4 @@
-import MuiButton from '@material-ui/core/Button';
-import styled, { keyframes } from 'styled-components';
+import styled from 'styled-components';
 
 export const HotelNotAvailableMessage = styled.div`
   height: 100%;
@@ -82,39 +81,4 @@ color: #3C3C3C;
   font-weight: 700;
   margin-bottom: 2px;
 }
-`;
-
-const fadeIn = keyframes`
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-`;
-
-export const RoomList = styled.ul`
-  display: flex;
-  flex-direction: row;
-  opacity: ${(props) => (props.isVisible ? '1' : '0')};
-  animation: ${fadeIn} 0.6s ease-in-out;
-  margin-bottom: 2rem;
-`;
-
-export const Title = styled.h5`
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 23px;
-  font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
-  color: #8e8e8e;
-  margin-bottom: 17px;
-  margin-top: 34px;
-  opacity: ${(props) => (props.isVisible ? '1' : '0')};
-  animation: ${fadeIn} 0.3s ease-in-out;
-`;
-
-export const BookingButton = styled(MuiButton)`
-  margin-top: 8px !important;
-  opacity: ${(props) => (props.isVisible ? '1' : '0')};
-  animation: ${fadeIn} 0.9s ease-in-out;
 `;
