+++ src/components/Hotels/index.js
@@ -9,23 +9,16 @@ import {
   HotelName,
   HotelCapacityInfo,
   HotelMainInfo,
-  RoomList,
-  Title,
-  BookingButton
 } from './HotelsStyles';
 import useGetHotels from '../../hooks/api/useGetHotels';
-import useSaveBooking from '../../hooks/api/useSaveBooking';
 import { useEffect, useState } from 'react';
 import { toast } from 'react-toastify';
 import hotelsHelpers from './helpers';
-import { Room } from '../Rooms';
 
-export default function HotelSelection({ ticket }) {
+export default function HotelSelection({ ticket, stateData }) {
   const { getHotels } = useGetHotels();
   const [hotels, setHotels] = useState(null);
-  const [selectedHotel, setSelectedHotel] = useState({});
-  const [selectedRoom, setSelectedRoom] = useState({});
-  const { saveBookingLoading, saveBooking } = useSaveBooking();
+  const { selectedHotel, setSelectedHotel } = stateData;
 
   useEffect(async () => {
     if ((ticket && ticket?.TicketType?.includesHotel && ticket?.status === 'PAID') && !hotels) {
@@ -44,17 +37,6 @@ export default function HotelSelection({ ticket }) {
       return;
     }
     setSelectedHotel({ id: data.id, rooms: data.Rooms });
-  };
-
-  const submitBooking = (data) => {
-    if (saveBookingLoading)
-      return;
-    if (!data) 
-      return toast('Por favor escolha um quarto.');
-    if (data.Booking.length === data.capacity)
-      return toast('Quarto sem vacâncias');
-
-    return saveBooking({ roomId: data.id });
   };
 
   return (
@@ -105,33 +87,6 @@ export default function HotelSelection({ ticket }) {
               </HotelOption>
             ))}
           </HotelsContainer>
-        </>
-      )}
-      {selectedHotel?.id && (
-        <>
-          <Title isVisible={selectedHotel?.id !== undefined}>
-            Ótima pedida! Agora escolha seu quarto
-          </Title>
-          <RoomList isVisible={selectedHotel?.id !== undefined}>
-            {(selectedHotel.rooms).map(data => {
-              return (
-                <Room
-                  data={data}
-                  key={data.id}
-                  selectedRoom={selectedRoom}
-                  setSelectedRoom={setSelectedRoom}
-                />
-              );
-            })}
-          </RoomList>
-          <BookingButton
-            variant='contained'
-            isVisible={selectedHotel?.id !== undefined}
-            onClick={() => submitBooking(selectedRoom.data)}
-            disabled={saveBookingLoading}
-          >
-            RESERVAR QUARTO
-          </BookingButton>
         </>
       )}
     </>
